daemon: factor K8s waitGroup synchronization into function
The same code was being used in multiple places, but for different Kubernetes
controllers, which makes refactoring it to be a function which is provided a
controller an easy, and more maintainable way of writing such code. Replace
existing equivalent code for NetworkPolicy and CiliumNetworkPolicy controllers,
and utilize it for Service, Ingress, and Endpoint controllers as well.

Signed-off by: Ian Vernon <ian@cilium.io>