refactor dump to userspace representation from daemon to lbmap
Converting from BPF key-value representation to the corresponding userspace
representations for the service and revNAT maps has several uses. Previously,
this was only done in the `SyncLB` function for the daemon. In an upcoming
commit, such functionality will be utilized elsewhere to synchronize the set of
services received by Kubernetes with the BPF maps. Factor out this functionality
into functions within the lbmap function accordingly. As an added bonus, make
some functions in the lbmap package private because they are not needed outside
of the lbmap package anymore after this refactor.

Signed-off by: Ian Vernon <ian@cilium.io>